Reorg mapping methods and fix build
