--- src/MadLearning/MadLearning.API.Application/Dtos/PersonModelDbDto.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
-
-namespace MadLearning.API.Application.Dtos
-{
-    public sealed class PersonModelDbDto
-    {
-        public string? Id { get; init; }
-
-        public string? FirstName { get; set; }
-
-        public string? LastName { get; set; }
-
-        public string? Email { get; set; }
-    }
-}
