--- src/MadLearning/MadLearning.API.Application/Mapping/EventModelDbDtoMappingExtensions.cs
@@ -1,27 +1,0 @@
-﻿using MadLearning.API.Application.Dtos;
-using MadLearning.API.Domain.Entities;
-using System;
-using System.Linq;
-
-namespace MadLearning.API.Application.Mapping
-{
-    public static class EventModelDbDtoMappingExtensions
-    {
-        public static EventModelDbDto ToDbDto(this EventModel? model)
-        {
-            if (model is null)
-                throw new ArgumentNullException(nameof(model));
-
-            return new EventModelDbDto
-            {
-                Id = model.Id,
-                Name = model.Name,
-                Description = model.Description,
-                ImageUrl = model.ImageUrl,
-                Owner = model.Owner.ToDbDto(),
-                Participants = model.Participants.ToDbDtos()?.ToList() ?? new (),
-                Time = model.Time,
-            };
-        }
-    }
-}
