+++ src/MadLearning/MadLearning.API.Infrastructure/DependencyInjection.cs
@@ -1,15 +1,10 @@
-﻿using MadLearning.API.Application.Dtos;
-using MadLearning.API.Application.Persistence;
+﻿using MadLearning.API.Application.Persistence;
 using MadLearning.API.Infrastructure.Configuration;
 using MadLearning.API.Infrastructure.Persistence;
 using MadLearning.API.Infrastructure.Services;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using MongoDB.Bson;
-using MongoDB.Bson.Serialization;
-using MongoDB.Bson.Serialization.IdGenerators;
-using MongoDB.Bson.Serialization.Serializers;
 
 namespace MadLearning.API.Infrastructure
 {
@@ -30,24 +25,6 @@ namespace MadLearning.API.Infrastructure
                 // Dependencies
                 services.AddSingleton<IEventRepository, EventRepository>();
                 services.AddSingleton<Application.Services.IIdGenerator, IdGenerator>();
-            }
-
-            {
-                // Register BSON class maps
-                BsonClassMap.RegisterClassMap<EventModelDbDto>(classMap =>
-                {
-                    classMap.AutoMap();
-                    classMap.MapIdProperty(c => c.Id)
-                        .SetIdGenerator(StringObjectIdGenerator.Instance)
-                        .SetSerializer(new StringSerializer(BsonType.ObjectId));
-                });
-                BsonClassMap.RegisterClassMap<PersonModelDbDto>(classMap =>
-                {
-                    classMap.AutoMap();
-                    classMap.MapIdProperty(c => c.Id)
-                        .SetIdGenerator(StringObjectIdGenerator.Instance)
-                        .SetSerializer(new StringSerializer(BsonType.ObjectId));
-                });
             }
 
             return services;
